cloud_bot/requirements.txt
gradio==4.20.0
pyTelegramBotAPI==4.16.1
llama-cpp-python==0.2.56
huggingface-hub
